+++ SprintMission05/src/pages/Items.jsx
@@ -1,6 +1,7 @@
-import Header from "../components/common/Header";
 import { createGlobalStyle } from "styled-components";
 import List from "../components/layout/List";
+import Header from "../components/layout/Header";
+import Nav from "../components/layout/Nav";
 
 const GlobalStyle = createGlobalStyle`
   * {
@@ -16,7 +17,16 @@ function Items() {
   return (
     <>
       <GlobalStyle />
-      <Header />
+      <Header
+        leftChild={
+          <div style={{ display: "flex", alignItems: "center" }}>
+            <img src="/src/assets/Logo.jpg" alt="Logo" /> <Nav />
+          </div>
+        }
+        rightChild={
+          <img src="/src/assets/ProfileIcon.jpg" alt="Profile Icon" />
+        }
+      />
       <List />
     </>
   );
